refactor: put custom JWT reused logic in core-kit
