+++ packages/toolkit/core-kit/package.json
@@ -41,6 +41,7 @@
     "node": "^20.9.0"
   },
   "dependencies": {
+    "@logto/connector-kit": "workspace:^3.0.0",
     "@logto/language-kit": "workspace:^1.1.0",
     "@logto/shared": "workspace:^3.1.0",
     "@silverhand/essentials": "^2.9.0",
