+++ packages/toolkit/core-kit/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "outDir": "lib"
   },
-  "include": ["src", "declaration"]
+  "include": ["src", "declaration", "../../schemas/src/types/logto-config/oidc-provider.ts"]
 }
