--- packages/toolkit/custom-jwt-kit/CHANGELOG.md
@@ -1,8 +1,0 @@
-# Change Log
-
-## 0.1.0
-
-Init commit of the Custom JWT Kit, including:
-
-1. General error handling
-2. Shared custom JWT utilities, such as running `getCustomJwtClaims` logic
