--- packages/toolkit/custom-jwt-kit/tsconfig.build.json
@@ -1,5 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "include": ["src"],
-  "exclude": ["src/**/*.test.ts"]
-}
