--- packages/toolkit/custom-jwt-kit/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "@silverhand/ts-config/tsconfig.base",
-  "compilerOptions": {
-    "outDir": "lib"
-  }
-}
